[patch] refactor code for view model
