+++ app/src/main/java/com/devspacecinenow/CineNowApp.kt
@@ -6,6 +6,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import com.devspacecinenow.detail.presentation.MovieDetailScreen
+import com.devspacecinenow.list.presentation.ui.MovieListScreen
 
 @Composable
 fun CineNowApp() {
