+++ app/src/main/java/com/devspacecinenow/common/model/MovieResponse.kt
@@ -1,4 +1,4 @@
-package com.devspacecinenow
+package com.devspacecinenow.common.model
 
 @kotlinx.serialization.Serializable
 data class MovieResponse(
