+++ app/src/main/java/com/devspacecinenow/detail/data/DetailService.kt
@@ -1,0 +1,12 @@
+package com.devspacecinenow.detail.data
+
+import com.devspacecinenow.common.model.MovieDto
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface DetailService {
+
+    @GET("{movie_id}?language=en-US")
+    fun getMovieById(@Path("movie_id") movie_id: String): Call<MovieDto>
+}
