+++ app/src/main/java/com/devspacecinenow/detail/presentation/MovieDetailScreen.kt
@@ -1,7 +1,6 @@
-package com.devspacecinenow
+package com.devspacecinenow.detail.presentation
 
 import android.util.Log
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
@@ -26,6 +25,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import coil.compose.AsyncImage
+import com.devspacecinenow.ApiService
+import com.devspacecinenow.common.model.MovieDto
+import com.devspacecinenow.common.data.RetrofitClient
 import com.devspacecinenow.ui.theme.CineNowTheme
 import retrofit2.Call
 import retrofit2.Callback
