+++ app/src/main/java/com/devspacecinenow/detail/presentation/MovieDetailViewModel.kt
@@ -1,0 +1,4 @@
+package com.devspacecinenow.detail.presentation
+
+class MovieDetailViewModel {
+}
