+++ app/src/main/java/com/devspacecinenow/list/data/ListService.kt
@@ -1,0 +1,20 @@
+package com.devspacecinenow.list.data
+
+import com.devspacecinenow.common.model.MovieResponse
+import retrofit2.Call
+import retrofit2.http.GET
+
+interface ListService {
+
+    @GET("now_playing?language=en-US&page=1")
+    fun getNowPlayingMovies(): Call<MovieResponse>
+
+    @GET("top_rated?language=en-US&page=1")
+    fun getTopRatedMovies(): Call<MovieResponse>
+
+    @GET("upcoming?language=en-US&page=1")
+    fun getUpcomingMovies(): Call<MovieResponse>
+
+    @GET("popular?language=en-US&page=1")
+    fun getPopularMovies(): Call<MovieResponse>
+}
