+++ app/src/main/java/com/devspacecinenow/list/presentation/MovieListViewModel.kt
@@ -1,0 +1,4 @@
+package com.devspacecinenow.list.presentation
+
+class MovieListViewModel {
+}
